
\name{Immunotherapy.Design-package}
\alias{Immunotherapy.Design}
\docType{package}
\title{
 Study design for immunotherapy clinical trials
}
\description{
 Estimate hazard rates and baseline hazards using an EM algorithm
}
\details{
 
}
\references{
 
}
\author{Zhenzhen Xu <Zhenzhen.Xu@fda.hhs.gov> and Bin Zhu <bin.zhu@nih.gov>}
\keyword{package}